Avoid caching of Matlab run command
The `run` command in Matlab caches the executed script, which causes the same lines to be executed again and again, instead of the selected ones.

This change clears the cached script, as per the Matlab documentation: https://www.mathworks.com/help/matlab/ref/run.html. Tested with Matlab R2017b.